feat: Use native.js file to make react native really optionnal
