+++ packages/cozy-client/src/authentication/mobile.js
@@ -6,10 +6,6 @@ import {
   isIOSApp
 } from 'cozy-device-helper'
 import { CordovaWindow } from '../types'
-//@ts-ignore
-import { InAppBrowser } from 'react-native-inappbrowser-reborn'
-//@ts-ignore
-import { Linking } from 'react-native'
 
 /**
  * @type {CordovaWindow}
@@ -95,60 +91,6 @@ const authenticateWithInAppBrowser = url => {
   })
 }
 
-/**
- * Opens a ReactNative InAppBrowsr
- * and resolves with the URL containing
- * the token
- *
- * @param {string} url
- * @returns {Promise}
- */
-
-export const authenticateWithReactNativeInAppBrowser = url => {
-  return new Promise((resolve, reject) => {
-    InAppBrowser.open(url, {
-      // iOS Properties
-      readerMode: false,
-      animated: true,
-      modalPresentationStyle: 'fullScreen',
-      modalTransitionStyle: 'coverVertical',
-      modalEnabled: true,
-      enableBarCollapsing: false,
-      // Android Properties
-      showTitle: true,
-      toolbarColor: '#6200EE',
-      secondaryToolbarColor: 'black',
-      enableUrlBarHiding: true,
-      enableDefaultShare: true,
-      forceCloseOnRedirection: false,
-      // Specify full animation resource identifier(package:anim/name)
-      // or only resource name(in case of animation bundled with app).
-      animations: {
-        startEnter: 'slide_in_right',
-        startExit: 'slide_out_left',
-        endEnter: 'slide_in_left',
-        endExit: 'slide_out_right'
-      }
-    })
-    const removeListener = () => {
-      Linking.removeEventListener('url', linkListener)
-    }
-
-    const linkListener = ({ url }) => {
-      const accessCode = /\?access_code=(.+)$/.test(url)
-      const state = /\?state=(.+)$/.test(url)
-
-      if (accessCode || state) {
-        resolve(url)
-        removeListener()
-        InAppBrowser.close()
-      }
-    }
-
-    Linking.addEventListener('url', linkListener)
-  })
-}
-
 export const authenticateWithCordova = async url => {
   if (isIOSApp() && (await hasSafariPlugin())) {
     return authenticateWithSafari(url)
@@ -165,7 +107,6 @@ export const authenticateWithCordova = async url => {
      * then get the "access_code" url and paste it in the prompt to let the
      * application initialize and redirect to other pages.
      */
-    console.log(url)
     return new Promise(resolve => {
       setTimeout(() => {
         const token = prompt('Paste the url here:')
@@ -175,13 +116,8 @@ export const authenticateWithCordova = async url => {
   }
 }
 
-const isReactNative = () => {
-  return typeof navigator != 'undefined' && navigator.product == 'ReactNative'
-}
 /**
  * Return the method to use for
  * authentication based on the
  */
-export const authFunction = isReactNative()
-  ? authenticateWithReactNativeInAppBrowser
-  : authenticateWithCordova
+export const authFunction = authenticateWithCordova
